frontend: router: Extract getRoutePath to own file
