+++ frontend/src/components/authchooser/index.tsx
@@ -26,7 +26,8 @@ import { getAppUrl } from '../../helpers/getAppUrl';
 import { getCluster, getClusterPrefixedPath } from '../../lib/cluster';
 import { useClustersConf } from '../../lib/k8s';
 import { testAuth } from '../../lib/k8s/api/v1/clusterApi';
-import { createRouteURL, getRoute, getRoutePath } from '../../lib/router';
+import { createRouteURL, getRoute } from '../../lib/router';
+import { getRoutePath } from '../../lib/router/getRoutePath';
 import { setConfig } from '../../redux/configSlice';
 import { ClusterDialog } from '../cluster/Chooser';
 import { DialogTitle } from '../common/Dialog';
