+++ frontend/src/components/common/AlertNotification.tsx
@@ -22,7 +22,8 @@ import { useTranslation } from 'react-i18next';
 import { matchPath, useLocation } from 'react-router-dom';
 import { getCluster } from '../../lib/cluster';
 import { testClusterHealth } from '../../lib/k8s/api/v1/clusterApi';
-import { getRoute, getRoutePath } from '../../lib/router';
+import { getRoute } from '../../lib/router';
+import { getRoutePath } from '../../lib/router/getRoutePath';
 
 // in ms
 const NETWORK_STATUS_CHECK_TIME = 5000;
